refactor(browser): split update match into update_media and update_remote handlers
